style: Cleaning the project structure
Every component is now broken down
into the principles of atomic design.
