+++ client/components/atoms/Fade/fade.module.scss
@@ -1,0 +1,11 @@
+.fade {
+  position: absolute;
+  left: 0%;
+  top: auto;
+  right: 0%;
+  bottom: 70px;
+  z-index: 2;
+  width: 100%;
+  height: 25px;
+  background-image: linear-gradient(180deg, hsla(0, 0%, 100%, 0), #fff);
+}
