+++ client/components/atoms/Fade/fade.tsx
@@ -1,0 +1,7 @@
+import styles from "./fade.module.scss"
+
+export default function Fade() {
+  return (
+    <div className={styles.fade}></div>
+  )
+}
