+++ client/components/atoms/Loader/loader.module.scss
@@ -1,0 +1,24 @@
+.loader {
+  position: relative;
+  display: inline-block;
+  max-width: 90%;
+  padding: 10px 20px 15px 20px;
+  border-radius: 4px;
+  background-color: #2962ff;
+  margin-top: 10px;
+  margin-bottom: 7px;
+}
+
+.loader:before {
+  content: url('./../../../public/svgs/bubble-corner.svg');
+  position: absolute;
+  top: calc(100% - 12px);
+  fill: blue;
+  left: 0;
+  height: 20px;
+  width: 20px;
+}
+
+.loaderDotsIcon {
+  height: 10px;
+}
