+++ client/components/atoms/Loader/loader.tsx
@@ -1,0 +1,9 @@
+import styles from "./loader.module.scss"
+
+export default function Loader() {
+  return (
+    <div className={styles.loader}>
+      <img src="/svgs/loading-dots-icon.svg" alt="Loading indicator icon" className={styles.loaderDotsIcon} />
+    </div>
+  )
+}
