--- client/components/atoms/fade.tsx
@@ -1,7 +1,0 @@
-import styles from "../../styles/Home.module.scss"
-
-export default function Fade() {
-  return (
-    <div className={styles.fadeBottom}></div>
-  )
-}
