+++ client/components/molecules/ChatMenu/chatMenu.module.scss
@@ -1,0 +1,87 @@
+.chatMenu {
+  position: absolute;
+  left: 0%;
+  top: 60px;
+  right: 0%;
+  bottom: auto;
+  z-index: 1;
+  width: 100%;
+  height: 450px;
+  padding-top: 30px;
+  padding-right: 40px;
+  padding-left: 40px;
+  flex-direction: column;
+  align-items: flex-end;
+  background-color: #111924;
+}
+
+.menuLink {
+  text-decoration: none;
+}
+
+.menuText {
+  margin-top: 0px;
+  margin-bottom: 0px;
+  padding-top: 10px;
+  padding-bottom: 10px;
+  transition: opacity 200ms ease;
+  font-family: Roboto, sans-serif;
+  color: #fff;
+  font-size: 18px;
+  line-height: 18px;
+  font-weight: 400;
+  text-align: right;
+  cursor: pointer;
+}
+
+.menuText:hover {
+  opacity: 0.75;
+}
+
+.copyrightContainer {
+  position: absolute;
+  left: 0%;
+  top: auto;
+  right: 0%;
+  bottom: 0%;
+  display: flex;
+  width: 100%;
+  height: auto;
+  margin-bottom: 60px;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  align-self: flex-start;
+}
+
+.copyrightText {
+  margin-bottom: 15px;
+  opacity: 0.5;
+  font-family: Roboto, sans-serif;
+  color: #fff;
+  line-height: 14px;
+  text-align: center;
+}
+
+.socialIconsContainer {
+  display: flex;
+  width: 100%;
+  justify-content: center;
+  align-items: center;
+}
+
+.socialIcon {
+  width: 25px;
+  margin-right: 5px;
+  margin-left: 5px;
+  opacity: 0.5;
+}
+
+.socialIconLink {
+  transition: opacity 200ms ease;
+  cursor: pointer;
+}
+
+.socialIconLink:hover {
+  opacity: 0.5;
+}
