+++ client/components/molecules/ChatMessage/chatMessage.tsx
@@ -1,5 +1,5 @@
 import { useRef, useEffect } from "react"
-import styles from "../../styles/Home.module.scss"
+import styles from "./chatMessage.module.scss"
 
 export default function ChatMessage(props) {
   const messagesEndRef = useRef(null)
@@ -12,8 +12,8 @@ export default function ChatMessage(props) {
     scrollToBottom()
   }, [props.messages]);
 
-
   // TODO Conditionally display the other message types as well
+
   return (
     <div ref={messagesEndRef} className={props.messageAuthor === "parola" ? `${styles.message}` : `${styles.message} ${styles.user}`}>
       <p className={styles.messageText}>{props.messageText}</p>
