+++ client/components/molecules/QuickReply/quickReply.module.scss
@@ -1,0 +1,26 @@
+.quickReply {
+  display: block;
+  margin: 2.5px 1.5px;
+  padding: 5px 10px;
+  border-style: solid;
+  border-width: 1px;
+  border-color: #2962ff;
+  border-radius: 20px;
+  transition: background-color 400ms ease;
+  font-family: Roboto, sans-serif;
+  color: #2962ff;
+  line-height: 14px;
+  text-align: center;
+  cursor: pointer;
+}
+
+.quickReply:hover {
+  background-color: #2962ff;
+  color: #fff;
+}
+
+.quickReplyText {
+  margin-bottom: 0px;
+  font-size: 13px;
+  line-height: 13px;
+}
