--- client/components/molecules/quickReply.tsx
@@ -1,12 +1,0 @@
-import styles from "../../styles/Home.module.scss"
-
-export default function QuickReply(props) {
-  return (
-  <div className={styles.quickReply} onClick={() => {
-    props.handleOnboardingInfo(props.text)
-    props.removeOnboardingHandler()
-    }}>
-    <p className={styles.quickReplyText}>{props.text}</p>
-  </div>
-  )
-}
