+++ client/components/organisms/OnboardingInfo/onboardingInfo.module.scss
@@ -1,0 +1,40 @@
+.onboardingInfo {
+  position: absolute;
+  display: flex;
+  left: 0%;
+  top: auto;
+  right: 0%;
+  bottom: 70px;
+  height: 530px;
+  padding-right: 40px;
+  padding-bottom: 90px;
+  padding-left: 40px;
+  flex-direction: column;
+  justify-content: flex-end;
+  border-radius: 8px;
+  background-image: linear-gradient(180deg, hsla(0, 0%, 100%, 0.35), #fff 64%);
+}
+
+.onboardingText {
+  margin-bottom: 0px;
+  font-family: Roboto, sans-serif;
+  font-size: 18px;
+  line-height: 22px;
+}
+
+.onboardingIcon {
+  position: absolute;
+  left: auto;
+  top: auto;
+  right: 0%;
+  bottom: 0%;
+  height: 130px;
+  transform: rotate(-6deg) translate(-29px, -1px);
+  z-index: 10;
+}
+
+@media screen and (max-width: 479px) {
+  .onboardingInfo {
+    border-radius: 0px;
+  }
+}
