+++ client/components/organisms/QuickReplies/quickReplies.module.scss
@@ -1,0 +1,14 @@
+.quickRepliesContainer {
+  position: absolute;
+  left: 0%;
+  top: auto;
+  right: 0%;
+  bottom: 70px;
+  display: flex;
+  width: 100%;
+  padding: 20px;
+  flex-direction: row;
+  justify-content: flex-end;
+  flex-wrap: wrap;
+  align-items: baseline;
+}
