+++ client/components/organisms/QuickReplies/quickReplies.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react"
-import styles from "../../styles/Home.module.scss"
-import QuickReply from "../molecules/quickReply"
+import styles from "./quickReplies.module.scss"
+import QuickReply from "../../molecules/QuickReply/quickReply"
 
 export default function QuickReplies(props) {
   const [quickRepliesText, setQuickRepliesText] = useState([
